Refactor GrowthChart: actualizar las importaciones de tipos para mejorar la organización del código y la claridad.
