+++ src/ui/growth-chart/hooks/useChartDataForGender.ts
@@ -1,5 +1,5 @@
 import { useMemo } from 'react';
-import type { ChartData } from '../types';
+import type { ChartData } from '../data-sets';
 
 export function useChartDataForGender(gender: string, chartData: ChartData = {}) {
   const chartDataForGender = useMemo(() => {
