+++ src/ui/growth-chart/hooks/usePatientBirthdateAndGender.ts
@@ -1,6 +1,6 @@
 import { restBaseUrl, openmrsFetch } from '@openmrs/esm-framework';
 import useSWRImmutable from 'swr/immutable';
-import { GenderCodes } from '../types';
+import { GenderCodes } from '../data-sets';
 
 export interface PatientInfo {
   uuid: string;
